2.0. audio: Audio pipeline is now in 32 bit.      decklink_consumer: Added support for 32 bit audio.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.0.0.2@1321 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

Former-commit-id: 6016fd682b267d2886babd51b1faa9cdd1812d54

+++ core/mixer/write_frame.cpp
@@ -37,7 +37,7 @@ struct write_frame::implementation
 	ogl_device*									ogl_;
 	std::vector<std::shared_ptr<host_buffer>>	buffers_;
 	std::vector<safe_ptr<device_buffer>>		textures_;
-	std::vector<int16_t>						audio_data_;
+	std::vector<int32_t>						audio_data_;
 	const core::pixel_format_desc				desc_;
 	const void*									tag_;
 	core::field_mode::type						mode_;
@@ -142,14 +142,14 @@ write_frame& write_frame::operator=(write_frame&& other)
 void write_frame::swap(write_frame& other){impl_.swap(other.impl_);}
 
 boost::iterator_range<uint8_t*> write_frame::image_data(size_t index){return impl_->image_data(index);}
-std::vector<int16_t>& write_frame::audio_data() { return impl_->audio_data_; }
+std::vector<int32_t>& write_frame::audio_data() { return impl_->audio_data_; }
 const boost::iterator_range<const uint8_t*> write_frame::image_data(size_t index) const
 {
 	return boost::iterator_range<const uint8_t*>(impl_->image_data(index).begin(), impl_->image_data(index).end());
 }
-const boost::iterator_range<const int16_t*> write_frame::audio_data() const
+const boost::iterator_range<const int32_t*> write_frame::audio_data() const
 {
-	return boost::iterator_range<const int16_t*>(impl_->audio_data_.data(), impl_->audio_data_.data() + impl_->audio_data_.size());
+	return boost::iterator_range<const int32_t*>(impl_->audio_data_.data(), impl_->audio_data_.data() + impl_->audio_data_.size());
 }
 const void* write_frame::tag() const {return impl_->tag_;}
 const core::pixel_format_desc& write_frame::get_pixel_format_desc() const{return impl_->desc_;}
